backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
channels==4.0.0
daphne==4.0.0
django-environ==0.11.2
stripe==7.11.0
Pillow==10.2.08